hooked up transactions properly on dashboard screen
moved widget to correct file
added param to function to get rid of duplicate function
added text to cards on dashboard if no transactions/goals
